+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 ### Release highlights
+* git: Introduce `jj git sync` for a `git pull --rebase`-like workflow across multiple bookmarks.
 
 ### Breaking changes